+++ canvas-jupyterhub-rewriter/vite.config.js
@@ -1,0 +1,19 @@
+import { defineConfig } from 'vite';
+import { resolve } from 'path';
+
+export default defineConfig({
+  build: {
+    outDir: 'dist',
+    emptyOutDir: true,
+    lib: {
+      entry: resolve(__dirname, 'src/main.js'),
+      formats: ['es'],
+      fileName: () => 'rewriter.js',
+    },
+    rollupOptions: {
+      output: {
+        inlineDynamicImports: true,
+      },
+    },
+  },
+});